Added NNUE + search improvements, new bench | bench 31790142

+++ engine/src/history/mod.rs
@@ -20,9 +20,9 @@ impl HistEntry {
     }
 }
 // Get the bonus and malus for history at a given depth.
-pub fn hist_delta(depth: usize) -> (i16, i16) {
-    let bonus = hist_bonus_max().min(hist_bonus_mult() * depth as i16 - hist_bonus_base());
-    let malus = hist_malus_max().min(hist_malus_mult() * depth as i16 - hist_malus_base());
+pub fn hist_delta(depth: i16) -> (i16, i16) {
+    let bonus = hist_bonus_max().min(hist_bonus_mult() * depth - hist_bonus_base());
+    let malus = hist_malus_max().min(hist_malus_mult() * depth - hist_malus_base());
 
     (bonus, malus)
 }
